fix restaurant problem

+++ backend/restaurant-service/src/controllers/resturant.controller.js
@@ -22,10 +22,9 @@ exports.createHotel = async (req, res) => {
     if (!hotelName || !hotelAddress || !location || !opentime) {
       return res.status(400).json({ message: "Missing required fields" });
     }
-    console.log("fuck you bitches");
-
+    
     const newHotel = new Hotel({
-      userID: userId, // Replace with dynamic user ID if needed
+      userID: userId,
       hotelName,
       hotelAddress,
       metaData,
@@ -160,9 +159,19 @@ exports.updateHotel = async (req, res) => {
 // Delete a hotel by ID
 exports.deleteHotel = async (req, res) => {
   try {
-    const deletedHotel = await Hotel.findByIdAndDelete(req.params.id);
-    if (!deletedHotel) return res.status(404).json({ message: "Hotel not found" });
-    res.status(200).json({ message: "Hotel deleted successfully" });
+    const hotel = await Hotel.findById(req.params.id);
+
+    if(!hotel){
+      return res.status(404).json({ message: "Hotel not found" });
+  1 }
+
+    if(hotel.userID !== req.user.userId){
+      return res.status(403).json({ message: "You are not authorized to delete this hotel" });
+    }
+
+    await Hotel.findByIdAndDelete(req.params.id);
+    res.status(200).json("Hotel deleted successfully");
+    
   } catch (error) {
     res.status(500).json({ error: error.message });
   }
